% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biscuiteerData.R
\docType{package}
\name{biscuiteerData-package}
\alias{biscuiteerData-package}
\alias{_PACKAGE}
\alias{biscuiteerData}
\title{Supplementary Datasets for Biscuiteer}
\description{
A handful of datasets to be used in biscuiteer
}
\examples{

  Zhou_solo_WCGW_inCommonPMDs.hg19 <-
      biscuiteerDataGet("Zhou_solo_WCGW_inCommonPMDs.hg19.rda")
  PMDs.hg19 <- biscuiteerDataGet("PMDs.hg19.rda")

}
\seealso{
Useful links:
\itemize{
\item \url{https://github.com/trichelab/biscuiteer}
\item Report bugs at
\url{https://github.com/trichelab/biscuiteer/issues}
}
}
\author{
Timothy J Triche Jr \email{Tim.Triche@vai.org},

Wanding Zhou \email{Wanding.Zhou@vai.org},

Ben Johnson \email{Ben.Johnson@vai.org},

Jacob Morrison \email{Jacob.Morrison@vai.org},

Lyong Heo \email{Lyong.Heo@vai.org}
}
\keyword{Biscuit}
\keyword{DNAMethylation}
\keyword{DataImport}
\keyword{EpigeneticsWorkflow}
\keyword{ExperimentData}
\keyword{MethylSeq}
