% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{biscuiteerDataCacheAll}
\alias{biscuiteerDataCacheAll}
\title{Cache all biscuiteer data}
\usage{
biscuiteerDataCacheAll(dateAdded = "all", showProgress = FALSE)
}
\arguments{
\item{dateAdded}{Version of the data (given by the date added), if "all"
then all dates will be cached (DEFAULT: "all")}

\item{showProgress}{Whether to show progress of download (DEFAULT: FALSE)}
}
\value{
\preformatted{         TRUE
}
}
\description{
Cache all biscuiteer data
}
\examples{

  biscuiteerDataCacheAll()

}
