% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{biscuiteerDataGet}
\alias{biscuiteerDataGet}
\title{Retrieve biscuiteer data}
\usage{
biscuiteerDataGet(title, dateAdded = "2019-09-25", verbose = FALSE)
}
\arguments{
\item{title}{Title of the data}

\item{dateAdded}{Version of the data (given by the date added)
(DEFAULT: "2019-09-25")}

\item{verbose}{Whether to output ExperimentHub message (DEFAULT: FALSE)}
}
\value{
\preformatted{      Data object
}
}
\description{
Retrieve biscuiteer data
}
\examples{

  wcgws <- biscuiteerDataGet("Zhou_solo_WCGW_inCommonPMDs.hg19.rda")

}
