% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{biscuiteerDataList}
\alias{biscuiteerDataList}
\title{List all biscuiteer data}
\usage{
biscuiteerDataList(dateAdded = "all")
}
\arguments{
\item{dateAdded}{Version of the data (given by the date added), if "all" then
all dates will be shown (DEFAULT: "all")}
}
\value{
\preformatted{     All titles from biscuiteer data
}
}
\description{
List all biscuiteer data
}
\examples{

  biscuiteerDataList()

}
