% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{biscuiteerDataListDates}
\alias{biscuiteerDataListDates}
\title{List all versions of biscuiteer data}
\usage{
biscuiteerDataListDates()
}
\value{
Sorted unique dates in biscuiteer data
}
\description{
List all versions of biscuiteer data
}
\examples{

  biscuiteerDataListDates()

}
