% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PMDs.hg38}
\alias{PMDs.hg38}
\title{PMDs.hg38}
\description{
Common PMD locations in hg38 genome
}
\details{
Source URL: http://zwdzwd.io/pmd/PMD_coordinates_hg38.bed.gz
Source type: BED file
Return type: GRanges
}
\keyword{data}
