% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Zhou_solo_WCGW_inCommonPMDs.hg19}
\alias{Zhou_solo_WCGW_inCommonPMDs.hg19}
\title{Zhou_solo_WCGW_inCommonPMDs.hg19}
\description{
Solo WCGWs living in hg19 common PMDs
}
\details{
Source URL: http://zwdzwd.io/pmd/solo_WCGW_inCommonPMDs_hg19.bed.gz
Source type: BED file
Return type: GRanges
}
\keyword{data}
